added error mapping to display errors when api is called unsuccessfully and beers are deleted that are referenced by batches

+++ src/main/java/com/sigvetl/twoDog/controller/BeerController.java
@@ -35,6 +35,9 @@ public class BeerController {
     @Autowired
     private BatchService batchService;
 
+    @Autowired
+    private ErrorController errorController;
+
     @PostMapping("/beers")
     public String addUpdateBeers(Authentication auth, BeerForm beerForm, BatchForm batchForm, Model model){
         if (this.beerService.beerExists(beerForm)){
@@ -68,15 +71,16 @@ public class BeerController {
                     this.beerService.trackLoggedInUserId(auth.getName());
                     this.beerService.createFromApi(beer);
                 }
-                System.out.println("beer: " +  beer.getName());
-                System.out.println("ibu: " +  beer.getIbu());
-                System.out.println("recipe link: " +  beer.getRecipeLink());
-                System.out.println("picture link: " +  beer.getPictureLink());
-                System.out.println("abv: " +  beer.getAbv());
-                System.out.println("ebc: " +  beer.getEbc());
+//                For error checking the api values
+//                System.out.println("beer: " +  beer.getName());
+//                System.out.println("ibu: " +  beer.getIbu());
+//                System.out.println("recipe link: " +  beer.getRecipeLink());
+//                System.out.println("picture link: " +  beer.getPictureLink());
+//                System.out.println("abv: " +  beer.getAbv());
+//                System.out.println("ebc: " +  beer.getEbc());
             }
         } catch(ConnectException connectException){
-            System.out.println("api is not live");
+            return this.errorController.error("Could not connect to API providing beers. Check that the server is running and try again.", model);
         }
 
             HomeController.updateHome(auth, model, this.beerService, this.batchService, this.userService);
@@ -85,11 +89,14 @@ public class BeerController {
 
     @GetMapping("/beers/delete/{beerid}")
     public String deleteBeer(@PathVariable("beerid") Integer beerId, Authentication auth, BeerForm beerForm, BatchForm batchForm, Model model){
-        this.beerService.deleteBeer(beerId);
-
-        HomeController.updateHome(auth, model, this.beerService, this.batchService, this.userService);
-
-        return "home";
+        if (batchService.getBatchReferencingBeer(beerId).size() == 0){
+            this.beerService.deleteBeer(beerId);
+            HomeController.updateHome(auth, model, this.beerService, this.batchService, this.userService);
+            return "home";
+        } else{
+            return this.errorController.error("One or more batches are referencing this beer. " +
+                    "You need to delete these before deleting this beer.", model);
+        }
     }
 
 }
